requirements.txt
fastapi
uvicorn
jinja2
requests
python-dotenv
pytest
